requirements.txt
vosk
pandas
xmltodict
requests
pydub
ffmpeg
python-dotenv
revChatGPT